logstash_checker.py: Change `--delay` argument type to float
Scap has always assumed that `logstash_checker.py` would accept a float
for `--delay`, but it seems it has only recently started calculating and
and passing a float which is causing `logstash_checker.py` to fail.

Let's change the argument type of `--delay` to a float to allow it to
accept either an int (backwards compatible) or a float (what scap
expects).

Bug: T308068

+++ modules/service/files/logstash_checker.py
@@ -87,7 +87,7 @@ class CheckService(object):
     """Shell class for checking services."""
 
     def __init__(self, host, service_name, logstash_host, user='', password='',
-                 verbose=False, delay=120, fail_threshold=2.0,
+                 verbose=False, delay=120.0, fail_threshold=2.0,
                  absolute_threshold=1.0):
         """Initialize the checker."""
         self.host = host
@@ -295,7 +295,7 @@ def main():
     parser.add_argument('--logstash-host',
                         default='https://logstash.wikimedia.org:9200',
                         help='The logstash host.')
-    parser.add_argument('--delay', default=120, type=int,
+    parser.add_argument('--delay', default=120.0, type=float,
                         help='Length of the delay (in seconds) between a '
                              'deploy & the call to this check script')
     parser.add_argument('--fail-threshold', type=float, default=2.0,
